feat: add debug mode

+++ src/utils/upload-profile.ts
@@ -1,15 +1,14 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
 import fs from 'fs-extra'
 import path from 'path'
 
-import { tmpRoot } from '../run'
 import { runCommand } from './exec'
 
 export const writeProfileToGitHubWithRetry = async (profilePath: string) => {
   const name = path.basename(profilePath)
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
   const { repoSetup } = require('../prepare/repo-setup')
-  // eslint-disable-next-line @typescript-eslint/no-var-requires
   const { actionInfo } = require('../prepare/action-info')
+  const { tmpRoot } = require('../run')
 
   const git = repoSetup({
     githubToken: process.env.PR_STATS_COMMENT_TOKEN!,
@@ -24,25 +23,29 @@ export const writeProfileToGitHubWithRetry = async (profilePath: string) => {
 
   const profile = path.join(profileRepoDir, name)
 
-  await fs.move(profilePath, profile)
-  await runCommand('git', ['status'], {
-    cwd: profileRepoDir
-  })
-  await runCommand('git', ['add', profile], {
-    cwd: profileRepoDir
-  })
-  await runCommand('git', [
-    '-c',
-    'user.name=speedy-benchmark-action',
-    '-c',
-    'user.email=github@users.noreply.github.com',
-    'commit',
-    '-m',
-    `add ${name} profile for ${actionInfo.prRef}`
-  ], {
-    cwd: profileRepoDir
-  })
-  await git.push('main', 'speedy-js/speedy-profiles', profileRepoDir)
+  if (fs.existsSync(profile)) {
+    console.log('profile exists, skipping upload', name)
+  } else {
+    await fs.move(profilePath, profile)
+    await runCommand('git', ['status'], {
+      cwd: profileRepoDir
+    })
+    await runCommand('git', ['add', profile], {
+      cwd: profileRepoDir
+    })
+    await runCommand('git', [
+      '-c',
+      'user.name=speedy-benchmark-action',
+      '-c',
+      'user.email=github@users.noreply.github.com',
+      'commit',
+      '-m',
+      `add ${name} profile for ${actionInfo.prRef}`
+    ], {
+      cwd: profileRepoDir
+    })
+    await git.push('main', 'speedy-js/speedy-profiles', profileRepoDir)
+  }
 
   // do some cleanups
   await fs.remove(profileRepoDir)
